work in progress: query
This is a work in progress on adding `query` command to query the
database.

I think this is a failed attempty because of the broad number of
encodings we are using and because often times expected data is
`[]byte`. Raw byte data is very hard to deal with in shell.