Launch branch update (#896)
* Feature/empty targets (#877)

* x/username: targets can be empty

* changelog: update

* x/username: fix tests

* unify query paths

Ensure all query paths follow the same naming convention.

resolve #849

* aswap: rename ReturnMsg (#879)

* Fixed failing scenario test

* x/gov: more query path unification

* Allow to use !nochangelog to mark pull request. (#883)

* Allow to use !nochangelog to mark pull request.

We use `#trivial` to allow merging a pull request without any
modyfications to the CHANGELOG file.

In addition, allow to use `!nochangelog` string to achieve the same
result.
`#trivial` is triggering GitHub's stupid autocompletion. It might also
be confusing to people from outside of the project what it means.

Do not search for any of those strings in the pull request title as we
never put it there.

* update README

* Implement cron functionality for messages.

This work is done to complete #671

A new extension `x/cron` is added. It allows to configure weave
application to be able to schedule messages for future execution.

Proposals created with `x/gov` are having their tally executed
automatically after the voting time is over. This is possible thanks to
`x/cron` extension.  This is not a breaking change because manual tally,
although pointless, is still possible.

Breaking change: `weave.Ticker` interface is updated.

* x/cron: move query registration to models.go

* x/cron: do not use sigs or multisig decorators

* x/cron: change return values of the Tick method

Instead of returning executed task IDs, Ticker.Tick method returns a
list of tags and a list of validator changes.

* x/cron: update error wrap message

* x/cron: ensure time rounding is always up

* x/cron: move declaration to the top

* x/cron: include execution time and block height in task info

* x/cache: improve how processing cache is manage

* attach block height to tested context

* x/gov: use 2s margin for running tally task

* cleanup go.sum

* x/cron: test task authentication

* x/cron: provide a limit of how many tasks can be processed

* x/gov: schedule tally task without authentication

* x/cron: test scheduler.Delete method

* bnsd: fix acceptance tests

* remove tmlibs import

tmplibs were merged into tendermint and must no longer be used. This
change fix the double protobuf message registration.

* weave: unused TickResult structure removed

* Update cron Ticker interface

Club together all return values of the `Ticker.Tick` method.

* x/cron: introduce a limit for the queue key searching

To avoid potential abuse of the task execution date for a huge amount of
tasks scheduled for the same time, introduce a limit for how many tries
there can be.

* commands/server/start: recover from panics to cleanup the socket (#885)

* commands/server/start: recover from panics to cleanup the socket

* server/start: log panic on recovery

* commands/server: fix type error

* Feature/escrow cleanup (#867)

* x/utils: remove a ctestify

* weavetest: introduced IsErr

* x/escrow: remove testify and unused tests

* util and escrow: convert to map[string]interface{} tests

* utils: fix a test

* escrow: address review comments

* escrow: assert.IsErr(t, nil -> assert.Nil(t...

* weavetest: fix PR comments and add tests to IsErr

* Simplified queries using helper function wrappers

* Simplify msgfee query

* Move useful setup helpers into tmtest

* Run real tendermint and bnsd server for clitests

* Properly set up tendermint nodes

* Update to use proper spam fee query

* WIP: pita to debug these scripts

* Fix signing test

* Fix up tests

* Better error message on empty accounts

* Moved RunBnsd, use TM_DEBUG to control output of tmlogs in tests

* Fix up test setup for bnsd/commands tests

* Fix up cmd_cash_test

* Build bnsd for travis

* Test helpers require smaller subset of testing.TB

* Use real tendermint/bnsd in bnscli tests, cmd_sigs_test updated

* submit test now does whole trip with real server

* Silence log messages on success in flag tests, fail on unexpected results

* cli tests don't use network, failures on those that require it

* Ensure proper shutdown of tendermint and bnsd at end of bnscli tests

* Kill with context.cancel, not defer cleanup() as well

* Last PR cleanup

* Use t.Logf rather than fmt.Println, added \n to mockAsserter

* Moved clitests that depend on network into README documentation

* Appease the linter

* x/gov: allow to change rule quorum

Update `UpdateElectionRuleMsg` and add `quorum` attribute to allow
overwriting existing election configuration.

`bnscli` updated to allow for the attribute setup.

resolve #869

* Update changelog.

* bnscli: improve fraction flag implementation

* security: add description and a pgp pub key (#889)

* security: add description and a pgp pub key

* security: refine phrase

* security: make pgp optional

* Remove `gov.TallyMsg` handler from message router. Provide a custom
router for the cron handler and register mostly necessary handlers,
inlcuding `gov.TallyMsg` handler.

This is a breaking change as describe in the `CHANGELOG`, because
`gov.TallyMsg` is no longer available to the user.

* remove gov.TallyMsg from bnsd transaction

* cmd/bnscli: remove tally

* Unify tests - all remaining packages (#888)

* Unify tests in app package

* Unifiy test in cmd package

* Unify tests remaining packages

* Go mod tidy

* Update Changelog

* Use map for table test

* Review feedback

* assign entity address at creation (#891)

* add self referencing attribute to entities

Add self referencing `address` attribute to entities `aswap.Swap`,
`escrow.Contract`, `distribution.Revenue`, `multisig.Contract` and
`paychan.PaymentChannel`.

Entity address can be build using that entity ID and few constants. For
convenience we include this value within the entity. This way the client
is not forced to implement address building logic.

resolve #644

* x/paychan: cleanup ID management

* fix tests

* update extensions to use orm.ModelBucket

All below extension are now using `orm.ModelBucket` to manage their
entities and an external sequecen instance to acquire ID.
- `x/aswap`
- `x/escrow`
- `x/distribution`
- `x/multisig`

* typo

* x/gov: add ElectionRule address and pregernate ID (#893)

* x/gov: add ElectionRule address and pregernate ID

* x/gov: address review comments and add tests

* changelog: update

* orm: fix linter error

* x/gov: add ElectionRule address test

* Update CHANGELOG.md (#895)

* Fix and polish flacky sigs/Decorator test (#897)