cmd/bnscli: add bench command
A new `bnscli bench` command was added that submits transactions from
the stdin using a single websocket connection. Transaction submission is
implemented in a way that allows measuring time from the first
transaction submission to the last transaction submission response.

`bnscli sign` was updated to allow manual sequence value choice instead
of fetching the information from the network.

An example script to benchmark submitting a transaction

    #!/bin/sh

    transactions=`mktemp 'lovenet-benchmark.XXXXXXXXXX'`
    tx_count=10
    username_base=`head /dev/urandom | tr -dc a-z | head -c 6 ; echo ''`
    # sequence must be precomputed and manually set
    sequence=XXX

    export BNSCLI_TM_ADDR=https://rpc.xxxxnet.iov.one:443

    # Create a batch transaction to put heavier load on the network.
    for i in `seq $tx_count`
    do
    	bnscli register-username -name "$username_base$i" -bc fake -addr addr \
		| bnscli with-fee -amount "100 IOV" \
		| bnscli sign -seq "$sequence"
    	sequence=$((sequence + 1))
    done > $transactions

    bnscli bench < $transactions

part of #931